Updating to run tests that are only compatibile on 1.5+
