+++ aws/request/http_request.go
@@ -20,7 +20,7 @@ func copyHTTPRequest(r *http.Request, body io.ReadCloser) *http.Request {
 		Method:        r.Method,
 		Proto:         r.Proto,
 		ContentLength: r.ContentLength,
-		// Cancel is deprecated in 1.7 and will be replaced with Context
+		// Cancel will be deprecated in 1.7 and will be replaced with Context
 		Cancel: r.Cancel,
 	}
 }
